requirements.txt
Jinja2==2.11.2
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
streamlit==0.66.0
tensorflow==2.3.0

